fixup compilation helper

+++ tools/inotify_make.sh
@@ -1,21 +1,23 @@
 #!/bin/bash
 echo "starting inotify listener on $(pwd)"
+ts=$(date +%s)
 # feed the inotify events into a while loop that creates
 # the variables 'date' 'time' 'dir' 'file' and 'event'
 inotifywait -mr --timefmt '%d/%m/%y %H:%M' --format '%T %w %f %e' \
 -e modify $(pwd) \
 | while read date time dir file event
 do
-    if [[ "$file" =~ \.go$ && "$dir" =~ "$(pwd)" ]]; then
-        echo --- $date $time ---
-        [ ! -z "$(pidof autograph)" ] && killall autograph
-        make install && make test
-        echo
-    fi
-    if [[ "$file" == autograph.yaml && "$dir" =~ "$(pwd)" ]]; then
-        echo --- $date $time ---
-        echo configuration changed, killing running autograph instance
-	    killall autograph
-        echo
+    # only run the loop once every 5 seconds.
+    cts=$(date +%s)
+    if [ "$((cts - ts))" -gt 5 ]; then
+        # execute compilation and tests
+        if [[ "$file" =~ \.go$ && "$dir" =~ "$(pwd)" ]]; then
+            echo --- $date $time ---
+            [ ! -z "$(pidof autograph)" ] && killall autograph
+            make all
+            echo
+            # move timestamp forward
+            ts=$cts
+        fi
     fi
 done
